+++ visualization/app/codeCharta/state/store/dynamicSettings/dynamicSettings.splitter.ts
@@ -7,7 +7,6 @@ import { splitEdgeMetricAction } from "./edgeMetric/edgeMetric.splitter"
 import { splitColorRangeAction } from "./colorRange/colorRange.splitter"
 import { splitMarginAction } from "./margin/margin.splitter"
 import { splitSearchPatternAction } from "./searchPattern/searchPattern.splitter"
-import { splitFocusedNodePathAction } from "./focusedNodePath/focusedNodePath.splitter"
 import { splitHeightMetricAction } from "./heightMetric/heightMetric.splitter"
 import { splitDistributionMetricAction } from "./distributionMetric/distributionMetric.splitter"
 import { splitColorMetricAction } from "./colorMetric/colorMetric.splitter"
@@ -44,10 +43,10 @@ export function splitDynamicSettingsActions(payload: RecursivePartial<DynamicSet
 		actions.push(splitSearchPatternAction(payload.searchPattern))
 	}
 
-	if (payload.focusedNodePath !== undefined) {
-		actions.push(splitFocusedNodePathAction(payload.focusedNodePath))
-	}
-
+	// is this used?
+	// if (payload.focusedNodePath !== undefined) {
+	// 	actions.push(splitFocusedNodePathAction(payload.focusedNodePath))
+	// }
 	if (payload.heightMetric !== undefined) {
 		actions.push(splitHeightMetricAction(payload.heightMetric))
 	}