+++ visualization/app/codeCharta/state/store/dynamicSettings/focusedNodePath/focusedNodePath.splitter.ts
@@ -1,7 +1,7 @@
 import { focusNode, unfocusNode } from "./focusedNodePath.actions"
 
-export function splitFocusedNodePathAction(payload: string) {
-	if (payload === "") {
+export function splitFocusedNodePathAction(payload?: string) {
+	if (!payload) {
 		return unfocusNode()
 	}
 	return focusNode(payload)