+++ analysis/import/SCMLogParser/src/test/kotlin/de/maibornwolff/codecharta/importer/scmlogparser/SCMLogProjectCreatorGoldenTest.kt
@@ -7,10 +7,9 @@ import de.maibornwolff.codecharta.importer.scmlogparser.parser.git.GitLogNumstat
 import de.maibornwolff.codecharta.importer.scmlogparser.parser.git.GitLogParserStrategy
 import de.maibornwolff.codecharta.importer.scmlogparser.parser.svn.SVNLogParserStrategy
 import de.maibornwolff.codecharta.model.Project
-import de.maibornwolff.codecharta.model.ProjectMatcher
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.serialization.ProjectSerializer
-import org.junit.Assert
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
@@ -71,14 +70,13 @@ class SCMLogProjectCreatorGoldenTest(
         val expectedProject = ProjectDeserializer.deserializeProject(ccjsonReader)
         val resource = this.javaClass.classLoader.getResource(logFilename)
         val logStream = Files.lines(Paths.get(resource!!.toURI()))
-
         // when
         val svnProject = svnSCMLogProjectCreator.parse(logStream)
         // This step is necessary because the comparison of the attribute map in MutableNode fails if the project is used directly;
         val svnProjectForComparison = serializeAndDeserializeProject(svnProject)
 
         // then
-        Assert.assertThat(svnProjectForComparison, ProjectMatcher.matchesProjectUpToVersion(expectedProject))
+        assertThat(svnProjectForComparison).usingRecursiveComparison().isEqualTo(expectedProject)
     }
 
     @Throws(IOException::class)