+++ CHANGELOG.md
@@ -22,11 +22,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Changed
 
 -   Replace all color pickers in the context of ongoing Angular migration [#2560](https://github.com/MaibornWolff/codecharta/pull/2560).
--   Remove suspicious metrics from custom map configurations [#2564](https://github.com/MaibornWolff/codecharta/pull/2564)
 
 ### Fixed 🐞
 
--   Fix showing names of custom configs properly [#2557](https://github.com/MaibornWolff/codecharta/pull/2557)
+-   Fix showing names of custom configs properly [#2557](https://github.com/MaibornWolff/codecharta/pull/2557).
+-   Fix minor memory leak in context menu for code map buildings [#2565](https://github.com/MaibornWolff/codecharta/issues/2565).
 
 ### Chore 👨‍💻 👩‍💻